Проектная работа 14

+++ controllers/cards.js
@@ -1,7 +1,7 @@
 const Card = require('../models/card');
 
 // Возвращает список всех карточек
-module.exports.getCards = (req, res) => {
+const getCards = (req, res) => {
   Card.find({})
     .populate('owner')
     .then((cards) => res.send({ data: cards }))
@@ -9,35 +9,44 @@ module.exports.getCards = (req, res) => {
 };
 
 // Создает новую карточку
-module.exports.createCard = (req, res) => {
+const createCard = (req, res) => {
   const { name, link, likes } = req.body;
   const userId = req.user._id;
 
-  // eslint-disable-next-line object-curly-newline
-  Card.create({ name, link, owner: userId, likes })
+  Card.create({
+    name,
+    link,
+    owner: userId,
+    likes,
+  })
     .then((card) => res.status(201).send({ data: card }))
-    // eslint-disable-next-line consistent-return
     .catch((err) => {
       if (err.name === 'ValidationError') {
         return res.status(400).send({ message: err.message });
       }
-      res.status(500).send({ message: err.message });
+      return res.status(500).send({ message: err.message });
     });
 };
 
 // Удаляет карточку с указанным id
-module.exports.deleteCard = (req, res, err) => {
-  Card.findByIdAndRemove(req.params.id)
-    // eslint-disable-next-line consistent-return
-    .then((result) => {
-      if (!result) {
-        return res.status(404).send({
-          message: 'Карточка с таким id отсутствует',
-          err: err.message,
-        });
+const deleteCard = (req, res) => {
+  Card.findById(req.params.id)
+    .then((card) => {
+      if (!card) {
+        return res.status(404).send({ message: `Карточка с id ${req.params.id} не найдена` });
       }
-      res.send({ data: result });
+      if (card.owner._id.toString() === req.user._id) {
+        return Card.findByIdAndRemove(req.params.id)
+          .then((result) => res.status(200).send({ message: `Карточка с id ${result._id} удалена` }))
+          .catch((err) => res.status(400).send({ message: err.message }));
+      }
+      return res.status(403).send({ message: 'Необходимо авторизоваться чтобы удалить карточку' });
     })
-    // eslint-disable-next-line no-shadow
     .catch((err) => res.status(500).send({ message: err.message }));
+};
+
+module.exports = {
+  deleteCard,
+  createCard,
+  getCards,
 };
